refactor: Split packagesettings module into a package
The `packagesettings` module has grown to over 1,200 lines of code. With
upcoming plans to improve resolver, downloader, and patching, the code
is going to grow even more.

Split the module into a package with several private modules. The public
interface of the `fromager.packagesettings` is mostly staying the same.
Only one monkey-patching case for testing had to be changed.
